requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
redis==5.0.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
gunicorn==21.2.0
Flask-Assets==2.1.0
cssmin==0.2.0
jsmin==3.0.1
typing-extensions==4.9.0
aiosmtplib==3.0.1
